+++ requirements.txt
@@ -1,12 +1,10 @@
-imageio==2.35.1
-numpy==2.0.2
-pillow==10.4.0
-protobuf==3.20
-scikit-image==0.24.0
-scipy==1.13.1
-tifffile==2024.8.30
+scikit-image
 tensorboard_logger
-torch==2.4.1
-torchvision==0.19.1
-segmentation-models-pytorch==0.3.4
-transformers==4.47.1
+torch
+torchvision
+transformers
+segmentation-models-pytorch
+wandb
+git+https://github.com/facebookresearch/sam2
+protobuf==3.20.*
+
